% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{formatter}
\alias{formatter}
\title{Format and pretty-print SSW forced alignment results without truncation}
\usage{
formatter(x, print = FALSE)
}
\arguments{
\item{x}{An object of class \code{ssw} containing the forced alignment results.}

\item{print}{Pretty-print the results? Default is \code{FALSE}.}
}
\value{
The formatted forced alignment results.
}
\description{
Format and pretty-print SSW forced alignment results without truncation
}
\examples{
\dontshow{if (is_installed_ssw_py()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
a <- force_align("ACTG", "TTTTCTGCCCCCACG")
b <- a |> formatter()
b
a |> formatter(print = TRUE)
\dontshow{\}) # examplesIf}
}
