% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_ssw_py}
\alias{install_ssw_py}
\title{Install ssw-py and its dependencies}
\usage{
install_ssw_py(
  ...,
  envname = "r-ssw-py",
  new_env = identical(envname, "r-ssw-py")
)
}
\arguments{
\item{...}{Other arguments passed to [reticulate::py_install()].}

\item{envname}{The name or full path of the environment in which
ssw-py is installed. Default is `r-ssw-py`.}

\item{new_env}{Logical. If `TRUE`, the specified Python environment
will be deleted and recreated if it already exists.
Defaults to `TRUE` only when using the default environment name.}
}
\value{
Invisibly returns `NULL`. Primarily used for its side effect
  of installing the Python package in the specified environment.
}
\description{
Install ssw-py and its dependencies
}
\examples{
\dontshow{if (is_installed_ssw_py()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
install_ssw_py()
\dontshow{\}) # examplesIf}
}
