% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.ssw}
\alias{print.ssw}
\title{Print SSW alignment results}
\usage{
\method{print}{ssw}(x, start_idx = 0L, ...)
}
\arguments{
\item{x}{An object of class \code{ssw}.}

\item{start_idx}{Index to start printing from.}

\item{...}{Additional parameters for \code{\link{print}} (not used).}
}
\description{
Print SSW alignment results
}
\examples{
\dontshow{if (is_installed_ssw_py()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
a <- align("ACGT", "TTTTACGTCCCCC")
a
\dontshow{\}) # examplesIf}
}
