% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{pyssw}
\alias{pyssw}
\title{Global reference to ssw-py}
\format{
An object of class \code{python.builtin.module} (inherits from \code{python.builtin.object}) of length 2.
}
\usage{
pyssw
}
\value{
ssw-py reference object
}
\description{
Global reference to ssw-py which will be initialized in \code{.onLoad}.
}
\keyword{datasets}
