% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{ssw-package}
\alias{ssw}
\alias{ssw-package}
\title{ssw: Striped Smith-Waterman Algorithm for Sequence Alignment using SIMD}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

R interface for SSW, a fast implementation of the Smith-Waterman algorithm for sequence alignment using SIMD.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nanx.me/ssw-r/}
  \item \url{https://github.com/nanxstats/ssw-r}
  \item Report bugs at \url{https://github.com/nanxstats/ssw-r/issues}
}

}
\author{
\strong{Maintainer}: Nan Xiao \email{me@nanx.me} (\href{https://orcid.org/0000-0002-0250-5673}{ORCID})

}
\keyword{internal}
