% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{ssw_py}
\alias{ssw_py}
\title{Global reference to ssw-py}
\format{
An object of class \code{python.builtin.module} (inherits from \code{python.builtin.object}) of length 0.
}
\usage{
ssw_py
}
\value{
ssw-py reference object
}
\description{
Global reference to ssw-py which will be initialized in \code{.onLoad}.
}
\keyword{datasets}
