% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{align}
\alias{align}
\title{Align a read to the reference with optional index offsetting}
\usage{
align(
  read = NULL,
  reference = NULL,
  gap_open = 3L,
  gap_extension = 1L,
  start_idx = 0L,
  end_idx = 0L,
  match_score = 2L,
  mismatch_penalty = 2L
)
}
\arguments{
\item{read}{The read, character string}

\item{reference}{The reference, character string}

\item{gap_open}{Penalty for gap_open}

\item{gap_extension}{Penalty for gap_extension}

\item{start_idx}{Index to start search}

\item{end_idx}{Index to end search (trying to avoid a target region).
The default value 0 means using the whole reference length.}

\item{match_score}{For scoring matches, integer}

\item{mismatch_penalty}{For scoring mismatches, integer}
}
\value{
A list containing the ssw aligner object and the alignment results.
}
\description{
Align a read to the reference with optional index offsetting
}
\examples{
\dontshow{if (is_installed_sswpy()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
a <- align("ACGT", "TTTTACGTCCCCC")
a
a$alignment$optimal_score
a$alignment$sub_optimal_score
\dontshow{\}) # examplesIf}
}
