% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/force_align.R
\name{force_align}
\alias{force_align}
\title{Forced alignment by increasing gap open penalty}
\usage{
force_align(read, reference, force_overhang = FALSE)
}
\arguments{
\item{read}{The read, character string}

\item{reference}{The reference, character string}

\item{force_overhang}{Make sure only one end overhangs?}
}
\value{
A list containing the input sequences, ssw aligner object,
and the alignment results.
}
\description{
Forced alignment by increasing gap open penalty
}
\examples{
# results are truncated
a <- force_align("ACTG", "TTTTCTGCCCCCACG")
a

# format the results
b <- a \%>\% formatter()
b

# print the formatted results directly
a \%>\% formatter(print = TRUE)
}
