% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/force_align.R
\name{force_align}
\alias{force_align}
\title{Forced alignment by increasing gap open penalty}
\usage{
force_align(
  read,
  reference,
  force_overhang = FALSE,
  match_score = 2L,
  mismatch_penalty = 2L
)
}
\arguments{
\item{read}{The read, character string}

\item{reference}{The reference, character string}

\item{force_overhang}{Make sure only one end overhangs?}

\item{match_score}{For scoring matches, integer}

\item{mismatch_penalty}{For scoring mismatches, integer}
}
\value{
A list containing the input sequences, ssw aligner object,
and the alignment results.
}
\description{
Forced alignment by increasing gap open penalty
}
\examples{
\dontshow{if (is_installed_sswpy()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Results are truncated
a <- force_align("ACTG", "TTTTCTGCCCCCACG")
a

# Format the results
b <- a |> formatter()
b

# Print the formatted results directly
a |> formatter(print = TRUE)
\dontshow{\}) # examplesIf}
}
