% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{formatter}
\alias{formatter}
\title{Format and pretty-print SSW forced alignment results without truncation}
\usage{
formatter(x, print = FALSE)
}
\arguments{
\item{x}{Forced alignment results. An object of class \code{ssw}.}

\item{print}{Pretty-print the results?}
}
\description{
Format and pretty-print SSW forced alignment results without truncation
}
\examples{
a <- force_align("ACTG", "TTTTCTGCCCCCACG")
b <- a \%>\% formatter()
b
a \%>\% formatter(print = TRUE)
}
