% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_installed.R
\name{is_installed_ssw_py}
\alias{is_installed_ssw_py}
\title{Is ssw-py installed?}
\usage{
is_installed_ssw_py()
}
\value{
\code{TRUE} if installed, \code{FALSE} if not.
}
\description{
Is ssw-py installed?
}
\examples{
is_installed_ssw_py()
}
